GUACAMOLE-408: exclude generated HTML files from rat validation
When building inside of the Docker container, some files with
generated HTML are being validated by RAT and failing due to the
lack of a license header. When building outside of Docker, this
problem does not occur. Not sure why these paths don't fail when
building outside of the container, and I don't know enough about
the plugin that generates these files to understand why it behaves
differently inside of the container. Nonetheless, I think these
excludes seem reasonable.