GUACAMOLE-470: Remove translations for "color-scheme" enum values.
Remove translations for the "color-scheme" enum values, because the
field type is now "TEXT" and not "ENUM" anymore. Changing the field type
to "TEXT" does have the disadvantage of losing localized enum choices.
Maybe it would be better to keep "color-scheme" an ENUM, and add a
"custom-color-scheme" TEXT field for customized color schemes?