GUACAMOLE-1807: Automatically check dependencies for updates and CVEs if "checkDependencies" is set to "true".
This can be done as part of the build:

    mvn -DcheckDependencies=true package

or separate from the build as a manual check:

    mvn -DcheckDependencies=true validate

Beware false positives and false negatives. Not all updates will
necessarily be caught be this, nor will all CVEs apply. The CVE matching
used is often a bit too fuzzy for its own good, but it gives an
excellent starting point for further manual checking.

The dependency update check will write each project's report to:

   .../target/dependency-update-report.txt

and the CVE check will write each project's report to:

   .../target/dependency-check-report.html

The relevant plugins WILL NOT genearte top-level, aggregate report.