GUACAMOLE-926: Reorder "Import" button to end to avoid fighting muscle memory.
Users that are used to the established Guacamole UI will be used to
clicking the leftmost button when attempting to create a new connection,
but the leftmost button has changed from "New Connection" to "Import".

This change restores that original button order, with "New Connection"
being on the far left.