GUACAMOLE-377: Revert to synchronous flush (asynchronous is slower).
The asynchronous flush mechanism leveraging requestAnimationFrame() does
not perform as well as the old synchronous flush. This appears to be due
to delays in when the browser actually allows the frame to proceed,
causing the client to lag behind.

The old synchronous flush mechanism does not suffer from such issues.